refactoring 'event' subcommands
